+++ tybotRoute/tiledeskChatbotPlugs/directives/DirAskGPT.js
@@ -15,6 +15,7 @@ class DirAskGPT {
       throw new Error('context object is mandatory');
     }
     this.context = context;
+    this.chatbot = this.context.chatbot;
     this.tdcache = this.context.tdcache;
     this.requestId = this.context.requestId;
     this.projectId = this.context.projectId;
@@ -97,11 +98,6 @@ class DirAskGPT {
 
     let key = await integrationService.getKeyFromIntegrations(this.projectId, 'openai', this.token);
     if (!key) {
-      winston.debug("(DirAskGPT) - Key not found in Integrations. Searching in kb settings...");
-      key = await this.getKeyFromKbSettings();
-    }
-
-    if (!key) {
       winston.debug("(DirAskGPT) - Retrieve public gptkey")
       key = process.env.GPTKEY;
       publicKey = true;
@@ -131,7 +127,10 @@ class DirAskGPT {
     let json = {
       question: filled_question,
       kbid: action.kbid,
-      gptkey: key
+      gptkey: key,
+      agent_id: this.chatbot?.bot.root_id || null,
+      id_project: this.projectId,
+      request_id: this.requestId
     };
     winston.debug("(DirAskGPT)DirAskGPT json:", json); 
 
@@ -244,29 +243,28 @@ class DirAskGPT {
     }
   }
 
-  async getKeyFromKbSettings() {
+  async checkQuoteAvailability() {
     return new Promise((resolve) => {
 
-      const KB_HTTPREQUEST = {
-        url: this.API_ENDPOINT + "/" + this.context.projectId + "/kbsettings",
+      const HTTPREQUEST = {
+        url: this.API_ENDPOINT + "/" + this.context.projectId + "/quotes/tokens",
         headers: {
           'Content-Type': 'application/json',
           'Authorization': 'JWT ' + this.context.token
         },
         method: "GET"
       }
-      winston.debug("(DirAskGPT) KB HttpRequest ", KB_HTTPREQUEST);
+      winston.debug("(DirAskGPT) check quote availability HttpRequest ", HTTPREQUEST);
 
       httpUtils.request(
-        KB_HTTPREQUEST, async (err, resbody) => {
+        HTTPREQUEST, async (err, resbody) => {
           if (err) {
-            winston.error("DirAskGPT Get kb settings error ", err?.response?.data);
-            resolve(null);
+            resolve(true)
           } else {
-            if (!resbody.gptkey) {
-              resolve(null);
+            if (resbody.isAvailable === true) {
+              resolve(true)
             } else {
-              resolve(resbody.gptkey);
+              resolve(false)
             }
           }
         }