+++ tybotRoute/tiledeskChatbotPlugs/directives/DirGptTask.js
@@ -124,12 +124,6 @@ class DirGptTask {
 
     let key = await integrationService.getKeyFromIntegrations(this.projectId, 'openai', this.token);
     if (!key) {
-      this.logger.native("[ChatGPT Task] Key not found in Integrations.");
-      winston.debug("(DirGptTask) - Key not found in Integrations. Searching in kb settings...");
-      key = await this.getKeyFromKbSettings();
-    }
-
-    if (!key) {
       this.logger.native("[ChatGPT Task] Retrieve shared gptkey.");
       winston.debug("(DirGptTask) - Retrieve public gptkey")
       key = process.env.GPTKEY;
@@ -317,36 +311,6 @@ class DirGptTask {
         await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, action.assignReplyTo, answer);
       }
     }
-  }
-
-  async getKeyFromKbSettings() {
-    return new Promise((resolve) => {
-
-      const KB_HTTPREQUEST = {
-        url: this.API_ENDPOINT + "/" + this.context.projectId + "/kbsettings",
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': 'JWT ' + this.context.token
-        },
-        method: "GET"
-      }
-      winston.debug("(DirGptTask) KB HttpRequest ", KB_HTTPREQUEST); 
-
-      httpUtils.request(
-        KB_HTTPREQUEST, async (err, resbody) => {
-          if (err) {
-            winston.error("(DirGptTask) Get KnowledgeBase err:", err.message);
-            resolve(null);
-          } else {
-            if (!resbody.gptkey) {
-              resolve(null);
-            } else {
-              resolve(resbody.gptkey);
-            }
-          }
-        }
-      )
-    })
   }
 
   async checkQuoteAvailability() {